% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.input.R
\name{AlleleFlip}
\alias{AlleleFlip}
\title{AlleleFlip}
\usage{
AlleleFlip(sumstats, snps, snpid = "snpid", only.a2 = F)
}
\arguments{
\item{sumstats}{A tibble, with at least columns snpid, a1, a2, beta, se, freq, pval.}

\item{snps}{A tibble, with at least columns snpid, a1, a2.}

\item{snpid}{Either a character indicating the common column name or,
in the case of two differently named columns,
a named character vector with the name of the common column in the first tibble as a name.}
}
\description{
Alleleflip joins a tibble with summary statistics and a tibble with SNP meta information,
by a common SNP id column, and aligns reference and alternative alleles to the meta tibble.
}
