% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar.conversion.R
\name{ConvertToPolar}
\alias{ConvertToPolar}
\title{ConvertToPolar}
\usage{
ConvertToPolar(
  dfnames,
  snpid,
  whiten = F,
  covsnps = c(),
  mahalanobis.threshold = 5,
  whitening.method = "ZCA-cor",
  LDcorrect = F,
  ld.path = "~/LDscores/eur_w_ld_chr/"
)
}
\arguments{
\item{snpid}{Either a character indicating the common column name or,
in the case of two differently named columns,
a named character vector with the name of the common column in the first tibble as a name.}

\item{covsnps}{A list of snpid's, indicating which rows (SNPs) should be included
for calculation of the covariance matrix. High confidence SNPs are recommended.}

\item{df1, df2}{tibble, with at least columns snpid, a1, a2, beta, se, freq, pval}
}
\description{
ConvertToPolar joins two tibbles with summary statistics by a common SNP id column,
with alleles already harmonized, for instance with AlleleFlip.
It then converts the original z-scores for trait 1 and 2 to polar coordinates, describing the angle theta with the x-axis (trait 1), and the distance r.
}
