% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.input.R
\name{ImportSNPs}
\alias{ImportSNPs}
\title{Import SNPs}
\usage{
ImportSNPs(
  file = "~/3.Athero-integration/mtag/mafgt0.1.infogt0.95.ngt10000.freqSElt0.05.snps",
  a1 = "a1",
  a2 = "a2"
)
}
\arguments{
\item{file}{A path for the SNP-file.}

\item{a1, a2}{The name of the column for the reference and alternative alleles.}
}
\description{
Imports a tibble with SNP meta information, and subsets to single nucleotide variants.
}
