% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{IsDitEenTest}
\alias{IsDitEenTest}
\title{Dit is een testfunctie}
\usage{
IsDitEenTest(dit)
}
\arguments{
\item{dit}{Het maakt niet uit wat dit is, dit is altijd een test}
}
\description{
Deze functie vertelt je of dit een test is.
}
\examples{
IsDitEenTest()
}
\keyword{test}
