% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.input.R
\name{LDCorrect}
\alias{LDCorrect}
\title{LDCorrect}
\usage{
LDCorrect(df, ld.path)
}
\arguments{
\item{df}{A tibble, with at least columns snpid and z-scores}

\item{ld.path}{the path to the LD scores files}
}
\description{
LDCorrect performs LD correction given a data-frame or tibble containing at least snpid's and z-scores (column names starting with "z"),
and a path to a LD scores file (as formatted by the LDSC software)
}
