% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar.conversion.R
\name{PolarCoords}
\alias{PolarCoords}
\title{PolarCoords}
\usage{
PolarCoords(inputdf, debug = F)
}
\arguments{
\item{inputdf}{vector of effect sizes (z-scores) of 1 SNP on p traits}
}
\value{
a list with the distance r, and the angle in radians, normalized to describe a full circle
}
\description{
Calculates the central angle between a SNP's position and the expected position under the null hypothesis of trait-specific effect.
The central angle is the angle between the vector from the origin of the p-dimensional sphere to the SNP, and the vector from the origin to the expected position.
Works per SNP (row-wise).
}
