% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{PvalueForAngle}
\alias{PvalueForAngle}
\title{PvalueForAngle}
\usage{
PvalueForAngle(
  angle.trans,
  r,
  p = 2,
  tol = 1e-20,
  kappa.file = "~/git/PolarMorphism/R/kappas.4foldtransform.Rda",
  debug = F
)
}
\arguments{
\item{angle.trans}{a vector of fourfold transformed angles}

\item{r}{a vector of distances}
}
\description{
PvalueForAngle performs p-value estimation using the von Mises distribution
}
