% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{PvalueForR}
\alias{PvalueForR}
\title{PvalueForR}
\usage{
PvalueForR(r, p = 2, exp.r = NA)
}
\arguments{
\item{r}{a vector of distances}

\item{p}{degrees of freedom (the number of original GWAS)}

\item{exp.r}{a vector of expected distances}
}
\description{
PvalueForR performs p-value estimation using the chi distribution
}
